+++ extras/tests/JsonDeserializer/destination_types.cpp
@@ -9,6 +9,9 @@
 
 #include "Allocators.hpp"
 
+using ArduinoJson::detail::sizeofArray;
+using ArduinoJson::detail::sizeofObject;
+
 TEST_CASE("deserializeJson(JsonDocument&)") {
   SpyingAllocator spy;
   JsonDocument doc(&spy);
@@ -23,6 +26,7 @@ TEST_CASE("deserializeJson(JsonDocument&)") {
                            Deallocate(sizeofPool()),
                            Deallocate(sizeofString("hello")),
                            Allocate(sizeofPool()),
+                           Reallocate(sizeofPool(), sizeofArray(1)),
                        });
 }
 
@@ -39,9 +43,11 @@ TEST_CASE("deserializeJson(JsonVariant)") {
 
     REQUIRE(err == DeserializationError::Ok);
     REQUIRE(doc.as<std::string>() == "[[42]]");
-    REQUIRE(spy.log() == AllocatorLog{
-                             Deallocate(sizeofString("hello")),
-                         });
+    REQUIRE(spy.log() ==
+            AllocatorLog{
+                Deallocate(sizeofString("hello")),
+                Reallocate(sizeofPool(), sizeofArray(1) + sizeofArray(1)),
+            });
   }
 
   SECTION("variant is unbound") {
@@ -64,17 +70,22 @@ TEST_CASE("deserializeJson(ElementProxy)") {
 
     REQUIRE(err == DeserializationError::Ok);
     REQUIRE(doc.as<std::string>() == "[[42]]");
-    REQUIRE(spy.log() == AllocatorLog{
-                             Deallocate(sizeofString("hello")),
-                         });
+    REQUIRE(spy.log() ==
+            AllocatorLog{
+                Deallocate(sizeofString("hello")),
+                Reallocate(sizeofPool(), sizeofArray(1) + sizeofArray(1)),
+            });
   }
 
-  SECTION("element must be created exists") {
+  SECTION("element must be created") {
     auto err = deserializeJson(doc[1], "[42]");
 
     REQUIRE(err == DeserializationError::Ok);
     REQUIRE(doc.as<std::string>() == "[\"hello\",[42]]");
-    REQUIRE(spy.log() == AllocatorLog{});
+    REQUIRE(spy.log() ==
+            AllocatorLog{
+                Reallocate(sizeofPool(), sizeofArray(2) + sizeofArray(1)),
+            });
   }
 }
 
@@ -89,9 +100,11 @@ TEST_CASE("deserializeJson(MemberProxy)") {
 
     REQUIRE(err == DeserializationError::Ok);
     REQUIRE(doc.as<std::string>() == "{\"hello\":[42]}");
-    REQUIRE(spy.log() == AllocatorLog{
-                             Deallocate(sizeofString("world")),
-                         });
+    REQUIRE(spy.log() ==
+            AllocatorLog{
+                Deallocate(sizeofString("world")),
+                Reallocate(sizeofPool(), sizeofObject(1) + sizeofArray(1)),
+            });
   }
 
   SECTION("member must be created exists") {
@@ -99,6 +112,9 @@ TEST_CASE("deserializeJson(MemberProxy)") {
 
     REQUIRE(err == DeserializationError::Ok);
     REQUIRE(doc.as<std::string>() == "{\"hello\":\"world\",\"value\":[42]}");
-    REQUIRE(spy.log() == AllocatorLog{});
+    REQUIRE(spy.log() ==
+            AllocatorLog{
+                Reallocate(sizeofPool(), sizeofObject(2) + sizeofArray(1)),
+            });
   }
 }